LA-106 Implement Room(@Entity, @Dao) for DownloadingTask table

+++ domain/src/main/java/com/linagora/android/linshare/domain/model/download/EnqueuedDownloadId.kt
@@ -1,6 +1,6 @@
 package com.linagora.android.linshare.domain.model.download
 
-data class EnqueuedDownloadId(val value: Int) {
+data class EnqueuedDownloadId(val value: Long) {
     init {
         require(value >= 0) { "enqueuedDownloadId must not be negative" }
     }
